Se finaliza la clase 15

+++ app/Http/Controllers/RepositoryController.php
@@ -7,6 +7,33 @@ use Illuminate\Http\Request;
 
 class RepositoryController extends Controller
 {
+
+    public function index(Request $request)
+    {
+        return view('repositories.index', [
+            'repositories' => $request->user()->repositories
+        ]);
+    }
+
+    
+    public function show(Request $request, Repository $repository) {
+
+        if($request->user()->id != $repository->user_id){
+            abort(403);
+        }
+        
+        return view('repositories.show', compact('repository')); 
+    }
+
+    public function edit(Request $request, Repository $repository) {
+
+        if($request->user()->id != $repository->user_id){
+            abort(403);
+        }
+        
+        return view('repositories.edit', compact('repository')); 
+    }
+
     public function store(Request $request){
 
         $request->validate([
@@ -37,7 +64,11 @@ class RepositoryController extends Controller
 
     }
 
-    public function destroy(Repository $repository) {
+    public function destroy(Request $request, Repository $repository) {
+
+        if($request->user()->id != $repository->user_id){
+            abort(403);
+        }
 
         $repository->delete();
